Update s0vz.c used double data type for aggmin

+++ s0vz.c
@@ -53,7 +53,7 @@ void signal_handler(int sig);
 void daemonize(char *rundir, char *pidfile);
 
 int pidFilehandle, vzport, len, running_handles, rc;
-float aggmin;
+double aggmin;
 
 const char *vzserver, *vzpath, *vzuuid[64];
